+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-04-25 22:48:35
-// Commit: feat(network): integrate custom telemetry logger in network
+// Updated at: 2026-04-26 10:07:38
+// Commit: feat(network): add bi-directional event bus handler for network
 
 class networkComponentState {
-  final String commitRef = '695';
-  final String lastMessage = 'feat(network): integrate custom telemetry logger in network';
-  final DateTime timestamp = DateTime.parse('2026-04-25T22:48:35+05:00');
+  final String commitRef = '699';
+  final String lastMessage = 'feat(network): add bi-directional event bus handler for network';
+  final DateTime timestamp = DateTime.parse('2026-04-26T10:07:38+05:00');
 }